dynaml-tensorflow:   -- Moved components away from tensorflow package object
