+++ dynaml-core/src/main/scala-2.11/io/github/mandar2812/dynaml/tensorflow/layers/MVTimeSeriesLoss.scala
@@ -16,7 +16,7 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 * */
-package io.github.mandar2812.dynaml.tensorflow.learn
+package io.github.mandar2812.dynaml.tensorflow.layers
 
 import org.platanios.tensorflow.api.learn.Mode
 import org.platanios.tensorflow.api.learn.layers.Loss
